fix: refactor some unit tests
