+++ packages/snap/tests/unit/Wallet.service.spec.ts
